fix: put key in a separate file, listed in .gitignore
